# api/openapi.yaml
openapi: 3.1.0
info:
  title: Ledger Core API
  version: 0.1.0
  description: Public REST API of the Ledger Core gateway.
servers:
  - url: http://localhost:8080/v1
security:
  - bearerAuth: []
paths:
...
  /transfers:
    post:
      summary: Create transfer
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [source_account_id, dest_account_id, amount, currency]
              properties:
                source_account_id: { type: string, format: uuid }
                dest_account_id: { type: string, format: uuid }
                amount: { type: string, pattern: "^[1-9][0-9]{0,18}$", description: Minor units as a decimal string }
                currency: { type: string }
                dest_currency: { type: string }
                description: { type: string }
      responses:
        "201": { description: Terminal status }
        "202": { description: In progress }
    get:
      summary: List transfers
      parameters:
        - { name: limit, in: query, schema: { type: integer } }
        - { name: cursor, in: query, schema: { type: string } }
      responses:
        "200": { description: OK }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 1, maxLength: 128 }
...